Standardize timestamp handling across all adapters to fix OpenAI Usage API 400 errors
Problem:
- BAEs adapter token tracking always showed 0 tokens
- OpenAI Usage API returned 400 Bad Request errors
- Root cause: BAEs used float timestamps (time.time()) with microseconds
- ChatDev and GHSpec already used int(time.time()) and worked correctly

Solution:
- Standardized BAEs adapter to use dual-timestamp pattern:
  * start_time = time.time() - float for precise duration calculation
  * start_timestamp = int(time.time()) - integer for API queries
  * duration = time.time() - start_time - preserves sub-second precision
  * end_timestamp = int(time.time()) - integer for API compatibility
- Added int() conversion safety layer in usage_reconciler.py
- Updated unit tests to expect integer timestamps

Changes:
- src/adapters/baes_adapter.py: Implement dual-timestamp pattern
- src/orchestrator/usage_reconciler.py: Add int() conversion for safety
- tests/unit/test_baes_adapter.py: Update timestamp type assertions

Results:
- All 79 unit tests passing
- Token reconciliation now working (25,607 in / 6,694 out tokens retrieved)
- All three adapters (BAEs, ChatDev, GHSpec) now consistent
- No more 400 API errors

Verification:
Run ID: 801758a3-dbda-4fc5-8d7f-fd1c916d7617
Status: Tokens successfully tracked and stored in manifest

+++ tests/unit/test_baes_adapter.py
@@ -313,8 +313,8 @@ class TestStepExecution:
         assert isinstance(result['duration_seconds'], float)
         assert isinstance(result['tokens_in'], int)
         assert isinstance(result['tokens_out'], int)
-        assert isinstance(result['start_timestamp'], float)
-        assert isinstance(result['end_timestamp'], float)
+        assert isinstance(result['start_timestamp'], int)  # Integer timestamps for API compatibility
+        assert isinstance(result['end_timestamp'], int)  # Integer timestamps for API compatibility
     
     @patch.object(BAeSAdapter, '_execute_kernel_request')
     def test_execute_step_token_placeholders(self, mock_execute_kernel, adapter):
